refactor: 체크박스 관리 useCheckbox 훅으로 분리 #238
